Develop directGuiderEnhancement

* Added setting to take into account side of pier as part of Meridian Flip.

* Logic to add a Direct Guider that sends PulseGuide commands to the mount without need for a guide camera.

* csproj cleanup

* Link UI parameter to Direct Guider logic

* Updating release notes

* Added direct guider attempts to connect to telescope when connected. This helps make Connect All work.

* Restore formatting

* Adapted to Stefan suggestions

* Changed logging level

* Small change to German label

* Random seed should not be re-generated each time - tends to send the telescope into similar direction each time

* Guide duration should allow decimal values

* Changed random angle selection logic to keep target deviation to a minimum

* Release notes

* Release Notes


Former-commit-id: 382b80b60d130a2b7b2e9d53a6679eb33d79e14e
Former-commit-id: e965f919fbcb6eb404ff1ae8e88cd7d62ebf1522

+++ NINA/Utility/Profile/GuiderSettings.cs
@@ -174,10 +174,10 @@ namespace NINA.Utility.Profile {
             }
         }
 
-        private int directGuideDuration;
+        private double directGuideDuration;
 
         [DataMember]
-        public int DirectGuideDuration {
+        public double DirectGuideDuration {
             get => directGuideDuration;
 
             set {
