+++ NINA/Utility/Profile/IGuiderSettings.cs
@@ -36,7 +36,7 @@ namespace NINA.Utility.Profile {
         int SettleTime { get; set; }
         double SettlePixels { get; set; }
         int SettleTimeout { get; set; }
-        int DirectGuideDuration { get; set; }
+        double DirectGuideDuration { get; set; }
         string PHD2Path { get; set; }
     }
 }
